requirements.txt
psycopg2-binary
pymongo
dotenv
pandas 
